+++ impl/src/lib.rs
@@ -57,7 +57,12 @@ mod mul_helpers;
 mod mul_like;
 #[cfg(feature = "not")]
 mod not_like;
-#[cfg(any(feature = "debug", feature = "display", feature = "from"))]
+#[cfg(any(
+    feature = "debug",
+    feature = "display",
+    feature = "from",
+    feature = "into",
+))]
 pub(crate) mod parsing;
 #[cfg(feature = "sum")]
 mod sum_like;